Update insert documentation

+++ CrudApp/Models/Product.cs
@@ -1,9 +1,10 @@
 namespace CrudApp.Models
 {
+    //This is the model for Product
     public class Product
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
     }
 }
